# libs/openchallenges/api-description/openapi/challenge.openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: OpenChallenges Challenge API
  description: Discover, explore, and contribute to open biomedical challenges.
  license:
    name: Apache 2.0
    url: https://github.com/Sage-Bionetworks/sage-monorepo/blob/main/LICENSE.txt
  contact:
    name: Support
    url: https://github.com/Sage-Bionetworks/sage-monorepo
  x-logo:
    url: https://dev.openchallenges.io/img/unsafe/logo/OpenChallenges-logo.png
servers:
  - url: http://localhost/v1
...
paths:
...
  /challengePlatforms:
    get:
      tags:
        - ChallengePlatform
      summary: List challenge platforms
      description: List challenge platforms
      operationId: listChallengePlatforms
      parameters:
        - $ref: '#/components/parameters/challengePlatformSearchQuery'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengePlatformsPage'
          description: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /challengePlatforms/{challengePlatformId}:
    parameters:
      - $ref: '#/components/parameters/challengePlatformId'
    get:
      tags:
        - ChallengePlatform
      summary: Get a challenge platform
      description: Returns the challenge platform identified by its unique ID
      operationId: getChallengePlatform
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengePlatform'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - ChallengePlatform
      summary: Delete a challenge platform
      description: |
        Deletes a challenge platform by its unique ID. This action is irreversible.
      operationId: deleteChallengePlatform
      security:
        - apiBearerAuth: []
      responses:
        '204':
          description: Deletion successful
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    apiBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: api_key
      description: API key obtained from /auth/login endpoint
  schemas:
...
    ChallengePlatformSlug:
      description: The slug of the challenge platform.
      type: string
      minLength: 3
      maxLength: 30
      pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
      example: example-challenge-platform
...
    PageMetadata:
      type: object
      description: The metadata of a page.
      properties:
        number:
          description: The page number.
          type: integer
          format: int32
          example: 99
        size:
          description: The number of items in a single page.
          type: integer
          format: int32
          example: 99
        totalElements:
          description: Total number of elements in the result set.
          type: integer
          format: int64
          example: 99
        totalPages:
          description: Total number of pages in the result set.
          type: integer
          format: int32
          example: 99
        hasNext:
          description: Returns if there is a next page.
          type: boolean
          example: true
        hasPrevious:
          description: Returns if there is a previous page.
          type: boolean
          example: true
      required:
        - number
        - size
        - totalElements
        - totalPages
        - hasNext
        - hasPrevious
...
    ChallengePlatformId:
      description: The unique identifier of a challenge platform.
      type: integer
      format: int64
      example: 1
    ChallengePlatformName:
      description: The name of the challenge platform.
      type: string
      minLength: 3
      maxLength: 30
...
    BasicError:
      type: object
      description: Problem details (tools.ietf.org/html/rfc7807)
      properties:
        title:
          type: string
          description: A human readable documentation for the problem type
        status:
          type: integer
          description: The HTTP status code
        detail:
          type: string
          description: A human readable explanation specific to this occurrence of the problem
        type:
          type: string
          description: An absolute URI that identifies the problem type
      required:
        - title
        - status
      x-java-class-annotations:
        - '@lombok.AllArgsConstructor'
        - '@lombok.Builder'
...
    ChallengePlatformSort:
      description: What to sort results by.
      type: string
      default: relevance
      enum:
        - name
        - relevance
    ChallengePlatformDirection:
      description: The direction to sort the results by.
      type: string
      nullable: true
      enum:
        - asc
        - desc
    ChallengePlatformSearchQuery:
      type: object
      description: A challenge platform search query.
      properties:
        pageNumber:
          description: The page number.
          type: integer
          format: int32
          default: 0
          minimum: 0
        pageSize:
          description: The number of items in a single page.
          type: integer
          format: int32
          default: 100
          minimum: 1
        sort:
          $ref: '#/components/schemas/ChallengePlatformSort'
        direction:
          $ref: '#/components/schemas/ChallengePlatformDirection'
        slugs:
          description: An array of challenge platform slugs used to filter the results.
          type: array
          items:
            $ref: '#/components/schemas/ChallengePlatformSlug'
        searchTerms:
          description: A string of search terms used to filter the results.
          type: string
          example: synapse
    ChallengePlatform:
      type: object
      description: A challenge platform
      properties:
        id:
          $ref: '#/components/schemas/ChallengePlatformId'
        slug:
          $ref: '#/components/schemas/ChallengePlatformSlug'
        name:
          $ref: '#/components/schemas/ChallengePlatformName'
        avatarKey:
          type: string
          example: logo/dream.png
        websiteUrl:
          type: string
          format: url
          example: https://example.com
        createdAt:
          type: string
          format: date-time
          example: '2022-07-04T22:19:11Z'
        updatedAt:
          type: string
          format: date-time
          example: '2022-07-04T22:19:11Z'
      required:
        - id
        - slug
        - name
        - avatarKey
        - websiteUrl
        - createdAt
        - updatedAt
    ChallengePlatformsPage:
      type: object
      description: A page of challenge platforms.
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            challengePlatforms:
              description: A list of challenge platforms.
              type: array
              items:
                $ref: '#/components/schemas/ChallengePlatform'
          required:
            - challengePlatforms
      x-java-class-annotations:
        - '@lombok.Builder'
...
  parameters:
...
    challengePlatformSearchQuery:
      name: challengePlatformSearchQuery
      description: The search query used to find challenge platforms.
      in: query
      schema:
        $ref: '#/components/schemas/ChallengePlatformSearchQuery'
    challengePlatformId:
      name: challengePlatformId
      in: path
      description: The unique identifier of the challenge platform.
      required: true
      schema:
        $ref: '#/components/schemas/ChallengePlatformId'
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    InternalServerError:
      description: The request cannot be fulfilled due to an unexpected server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    NotFound:
      description: The specified resource was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    Forbidden:
      description: The user does not have the permission to perform this action
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
...